Using a low-pass filter to smooth the compass data
git-svn-id: https://android-compass.googlecode.com/svn/trunk@15 4f6d0458-8dd9-cdf8-1905-fec939af8165

+++ src/com/jwetherell/compass/data/GlobalData.java
@@ -8,13 +8,13 @@ package com.jwetherell.compass.data;
  */
 public abstract class GlobalData {
 	private static final Object lock = new Object();
-    private static int bearing = 0;
+    private static float bearing = 0;
 
     /**
      * Set the bearing.
      * @param bearing int representing the current bearing.
      */
-    public static void setBearing(int bearing) {
+    public static void setBearing(float bearing) {
     	synchronized (lock) {
     		GlobalData.bearing = bearing;
     	}
@@ -24,7 +24,7 @@ public abstract class GlobalData {
      * Get the bearing.
      * @return int representing the bearing.
      */
-    public static int getBearing() {
+    public static float getBearing() {
     	synchronized (lock) {
     		return bearing;
     	}
